refac: Move createFinding to a specific file
