+++ Balancer-Bots/Swap-FlashLoan-Fee-Change/src/utils.ts
@@ -1,22 +1,5 @@
-import { ethers, Finding, FindingSeverity, FindingType } from "forta-agent";
-
 export type AgentConfig = Record<number, NetworkData>;
 
 export interface NetworkData {
   protocolFeesCollectorAddress: string;
-}
-
-export function createFinding(feeFrom: "FlashLoan" | "Swap", newFee: ethers.BigNumber): Finding {
-  return Finding.from({
-    name: `${feeFrom} fee changed`,
-    description: `A ${feeFrom} fee percentage change was detected`,
-    alertId: feeFrom === "FlashLoan" ? "BAL-1-1" : "BAL-1-2",
-    severity: FindingSeverity.Info,
-    type: FindingType.Info,
-    protocol: "Balancer",
-    metadata: {
-      // the fee value is a 18-decimal fixed point number, so by shifting left 16 places the result is the actual %
-      newFeePercentage: ethers.utils.formatUnits(newFee, 16),
-    },
-  });
 }
